requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
boto3
openai
tiktoken
faiss-cpu
langchain-community
langchain_google_genai